refactor(cli): extract provider commands
